moved action enum to new file
